requirements.txt
aes-pkcs5==1.0.1
asgiref==3.6.0
babel==2.16.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==39.0.1
defusedxml==0.7.1
Django==4.0
django-cors-headers==3.13.0
django-crontab==0.7.1
django-environ==0.11.2
django-jazzmin==2.6.0
djangorestframework==3.14.0
djangorestframework-xml==2.0.0
et-xmlfile==1.1.0
greenlet==3.0.3
gunicorn==20.1.0
idna==3.4
mysql-connector-python==9.0.0
mysqlclient==2.1.1
numpy==1.25.1
openpyxl==3.1.2
packaging==23.1
pandas==2.0.3
Pillow==9.5.0
pycparser==2.21
PyMySQL==1.1.1
pytesseract==0.3.10
python-dateutil==2.8.2
pytz==2022.7.1
requests==2.28.2
six==1.16.0
SQLAlchemy==2.0.32
sqlparse==0.4.3
typing_extensions==4.12.2
tzdata==2023.3
urllib3==1.26.15
xlrd==2.0.1
xmltodict==0.13.0
